music-recommendation-service/requirements.txt
pandas
pymongo
scikit-surprise
flask
python-dotenv
gunicorn